WIP: snackbar rendering inside hooks
